+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './hooks';
 export { useOnClickOutside } from './useOnClickOutside';
 export { useIsHovering } from './useIsHovering';
+export { useLaggedState } from './useLaggyState';